firstCheck

+++ function/changePassword.php
@@ -3,31 +3,42 @@
     include('verifieRegistration.php');
     function passwordResendVerifie($token,$email,$password,$confirmPassword,$PDO){
         try{
-            $messageSend = array();
-            if(!ft_verifie_password($password,$confirmPassword)){
-                $statement = "UPDATE `users` SET passwrd = ? WHERE email = ?";
-                $password = hash('whirlpool',$password);
-                $field = array($password,$email);
-                $PDO->statementPDO($statement,$field,0);
-                $field = array($token,$email);
-                $statement2 = "DELETE FROM `password_reset` WHERE token = ? AND EMAIL = ?";
-                $PDO->statementPDO($statement2,$field,0);
-                $valid = "your password have been reseted";
+            $statement = "SELECT * FROM `password_reset` WHERE token = ? AND email = ?";
+            $field = array($_GET['token'],$_GET['email']);            
+            if(!$PDO->statementPDO($statement,$field)){
+                $messageSend = array();
+                if(!ft_verifie_password($password,$confirmPassword)){
+                    $statement = "UPDATE `users` SET passwrd = ? WHERE email = ?";
+                    $password = hash('whirlpool',$password);
+                    $field = array($password,$email);
+                    $PDO->statementPDO($statement,$field,0);
+                    $field = array($token,$email);
+                    $statement2 = "DELETE FROM `password_reset` WHERE token = ? AND EMAIL = ?";
+                    $PDO->statementPDO($statement2,$field,0);
+                    $valid = "your password have been reseted";
+                }
+                else
+                    $error = ft_verifie_password($password,$confirmPassword); 
+                if(isset($valid) || isset($error)){
+                    if(!isset($valid))
+                        $valid = '';
+                    if(!isset($error))
+                        $error = '';
+                    echo json_encode($messageSend = array(
+                        'valid' => $valid,
+                        'error' => $error,
+                    ));            
+                }                 
             }
             else
-                $error = ft_verifie_password($password,$confirmPassword); 
-            if($valid || $error){
-                echo json_encode($messageSend = array(
-                    'valid' => $valid,
-                    'error' => $error,
-                ));            
-            }            
+                echo "";
+           
         }
         catch(Exeption $e){
             
         }
     }
-    if($_GET['resetPassword']&&$_GET['password']&&$_GET['email']&&$_GET['confirmPassword']&&$_GET['token']){
+    if(isset($_GET['resetPassword'])&&isset($_GET['password'])&&isset($_GET['email'])&&isset($_GET['confirmPassword'])&&isset($_GET['token'])){
         passwordResendVerifie($_GET['token'],$_GET['email'],$_GET['password'],$_GET['confirmPassword'],$PDO);
     }
 ?>
